[delete] etc

+++ src/main/java/com/sku_sku/MatchPrediction/dto/PredictionStatisticsRes.java
@@ -1,4 +1,4 @@
 package com.sku_sku.MatchPrediction.dto;
 
-public record PredictionStatisticsRes(String predictionResult, double percentage) {
+public record PredictionStatisticsRes(String predictionResult, int percentage) {
 }
